3408: Move image spanning to utils; tweak gui a bit (looks like status)
