+++ app/src/main/res/drawable/help_message_background.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+        android:bottom="1dp"
+        android:left="-1dp"
+        android:right="-1dp"
+        android:top="1dp">
+
+        <shape android:shape="rectangle">
+            <solid android:color="?attr/colorBackgroundAccent"/>
+            <stroke android:width="1dp" android:color="?attr/dividerColor"/>
+        </shape>
+    </item>
+</layer-list>
